+++ examples/cross_module/app/modules/orders/module.rb
@@ -1,0 +1,4 @@
+Shaolin.module "orders" do
+  commands_handled "Orders::Commands::PlaceOrder"
+  events_published "orders.order_placed"
+end